backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
email-validator
PyJWT
passlib[bcrypt]
python-dotenv
python-multipart
pandas
openpyxl
groq
google-generativeai
